Group modify commits by timestamp
Any surround commits that are modify a file (ie no branch operation) that
occur at the same timestamp, author, and branch are combined in the
export phase to appear as one commit.

Note timestamps that we get out of the sscm cli app only give us a
granularity of minutes, so it is possible (though unlikely) we
accidently group commits that are really two seperate commits.

Signed-off-by: Dalton Verhagen <verhagen.d@pg.com>